// log-manager-front/src/lib/api/query/queryKey.ts
export enum QUERY_KEY {
  ACCOUNT_GET_USERS = 'ACCOUNT_GET_USERS',
  HISTORY_GET_LIST = 'HISTORY_GET_LIST',
  CONFIGURE_GET_HOST = 'CONFIGURE_GET_HOST',
  CONFIGURE_GET_SITES = 'CONFIGURE_GET_SITES',

  CONFIGURE_GET_SITE_NAMES = 'CONFIGURE_GET_SITE_NAMES',
  CONFIGURE_GET_SITE_NAMES_NOT_ADDED = 'CONFIGURE_GET_SITE_NAMES_NOT_ADDED',
  STATUS_GET_LOCALS = 'STATUS_GET_LOCALS',
  STATUS_GET_REMOTES = 'STATUS_GET_REMOTES',
  JOB_REMOTE_GET_PLANS = 'JOB_REMOTE_GET_PLANS',
  JOB_REMOTE_GET_JOBS = 'JOB_REMOTE_GET_JOBS',

  ADDRESS_GET_GROUPS_EMAILS = 'ADDRESS_GET_GROUPS_EMAILS',
  ADDRESS_GET_GROUPS = 'ADDRESS_GET_GROUPS',
  ADDRESS_GET_EMAILS = 'ADDRESS_GET_EMAILS',
  ADDRESS_GET_EMAILS_BY_GROUP = 'ADDRESS_GET_EMAILS_BY_GROUP',
  ADDRESS_GET_GROUPS_IN_EMAIL = 'ADDRESS_GET_GROUPS_IN_EMAIL',

  RULES_CRAS_GET_INFO_LIST = 'RULES_CRAS_GET_INFO_LIST',
  RULES_CRAS_GET_SITE_INFO_LIST = 'RULES_CRAS_GET_SITE_INFO_LIST',

  RULES_CRAS_GET_MANUAL_CREATE = 'RULES_CRAS_GET_MANUAL_CREATE',
  RULES_CRAS_GET_MANUAL_CREATE_DETAIL = 'RULES_CRAS_GET_MANUAL_CREATE_DETAIL',

  RULES_CRAS_GET_MANUAL_CREATE_TARGET_TABLE_LIST = 'RULES_CRAS_GET_MANUAL_CREATE_TARGET_TABLE_LIST',
  RULES_CRAS_GET_MANUAL_CREATE_TARGET_COLUMN_LIST = 'RULES_CRAS_GET_MANUAL_CREATE_TARGET_COLUMN_LIST',

  RULES_CRAS_GET_MANUAL_JUDGE = 'RULES_CRAS_GET_MANUAL_JUDGE',
  RULES_CRAS_GET_MANUAL_JUDGE_DETAIL = 'RULES_CRAS_GET_MANUAL_JUDGE_DETAIL',

  RULES_CRAS_GET_MANUAL_CREATE_OPTION = 'RULES_CRAS_GET_MANUAL_CREATE_OPTION',
  RULES_CRAS_GET_MANUAL_JUDGE_OPTION = 'RULES_CRAS_GET_MANUAL_JUDGE_OPTION',
}